create query subpackage
